+++ grails-datastore-gorm-hibernate4/src/main/groovy/org/grails/orm/hibernate/HibernateSession.java
@@ -83,7 +83,7 @@ public class HibernateSession extends AbstractHibernateSession {
      * @param criteria The criteria
      * @return The total number of records deleted
      */
-    public int deleteAll(final QueryableCriteria criteria) {
+    public long deleteAll(final QueryableCriteria criteria) {
         return getHibernateTemplate().execute(new GrailsHibernateTemplate.HibernateCallback<Integer>() {
             public Integer doInHibernate(Session session) throws HibernateException, SQLException {
                 JpaQueryBuilder builder = new JpaQueryBuilder(criteria);
